% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplot.R
\docType{package}
\name{fplot-package}
\alias{fplot}
\alias{fplot-package}
\title{Aggregate/conditional graphs and automatic layout using formulas}
\description{
\pkg{fplot} provides automatic plotting of common graphs (distributions, lines, bar plots and boxplots). The syntax uses formulas, allowing aggregate/conditional/weighted graphs with minimum efforts. The many arguments are automatically adjusted to the data in order to provide the nicest and most meaningful graphs.
}
\details{
The core functions are: \code{\link[fplot]{plot_distr}} to draw distributions, \code{\link[fplot]{plot_lines}} to represent the (usually temporal) evolution of variables, \code{\link[fplot]{plot_box}} for conditional boxplots, and \code{\link[fplot]{plot_bar}} for aggregate bar plots.
}
\author{
\strong{Maintainer}: Laurent Berge \email{laurent.berge@uni.lu}

}
