% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_distr}
\alias{plot_distr}
\title{Plot distributions, possibly conditional}
\usage{
plot_distr(fml, base, moderator, weight, maxFirst, toLog, maxBins, bin,
  legend_options = list(), onTop, yaxisShow = TRUE, col,
  outCol = "black", mod.method = "splitWithin", labels.tilted,
  addOther, plot = TRUE, sep, centered = TRUE, weight.fun,
  dict = getFplot_dict(), mod.title, labels.angle, cex.axis,
  trunc = 20, trunc.method = "auto", ...)
}
\arguments{
\item{fml}{A formula or a vector. If a formula, it must be of the type: \code{var ~ moderator | weight}. If there are no moderator nor weights, you can use directly a vector, or use \code{fml = var ~ 1}. To use weights and no moderator, use \code{fml = var ~ 1 | weight}. See examples.}

\item{base}{A data.frame: data base containing the variables in the formula.}

\item{moderator}{Optional, only if argument \code{fml} is a vector. A vector of moderators.}

\item{weight}{Optional, only if argument \code{fml} is a vector. A vector of (positive) weights.}

\item{maxFirst}{Logical: should the first elements displayed be the most frequent? By default this is the case except for numeric values put to log or to integers.}

\item{toLog}{Logical, only used when the data is numeric. If \code{TRUE}, then the data is put to logarithm beforehand. By default numeric values are put to log if the log variation exceeds 3.}

\item{maxBins}{Maximum number of items displayed. The default depends on the number of moderator cases. When there is no moderator, the default is 15, augmented to 20 if there are less than 20 cases.}

\item{bin}{Only used for numeric values. If provided, it creates bins of observations of size \code{bin}. It creates bins by default for numeric non-integer data.}

\item{legend_options}{A list. Other options to be passed to \code{legend} which concerns the legend for the moderator.}

\item{onTop}{What to display on the top of the bars. Can be equal to "frac" (for shares), "nb" or "none". The default depends on the type of the plot.}

\item{yaxisShow}{Whether the y-axis should be displayed, default is \code{TRUE}.}

\item{col}{A vector of colors, default is close to paired. You can also use \dQuote{set1} or \dQuote{paired}.}

\item{outCol}{Outer color of the bars. Defaults is \code{"black"}.}

\item{mod.method}{A character scalar: either \dQuote{splitWithin}, the default, \dQuote{splitTotal}, \dQuote{within}, or \dQuote{total}. This is only used when there is more than one moderator. If within: the bars represent the distribution within each moderator class; if total, the heights of the bar represent the share in the total distribution. If split: there is one separate histogram for each moderator case.}

\item{labels.tilted}{Whether there should be tilted labels. Default is \code{FALSE} except when the data is split by moderators (see \code{mod.method}).}

\item{addOther}{Logical. Should there be a last column counting for the observations not displayed? Default is \code{TRUE} except when the data is split.}

\item{plot}{Logical, default is \code{TRUE}. If \code{FALSE} nothing is plotted, only the data is returned.}

\item{sep}{Positive number. The separation space between the bars. The scale depends on the type of graph.}

\item{centered}{Logical, default is \code{TRUE}. For numeric data only and when \code{maxFirst=FALSE}, whether the histogram should be centered on the mode.}

\item{weight.fun}{A function, by default it is \code{sum}. Aggregate function to be applied to the weight with respect to variable and the moderator. See examples.}

\item{dict}{A dictionnary to rename the variables names in the axes and legend. Should be a named vector. By default it s the value of \code{getFplot_dict()}, which you can set with the function \code{\link[fplot]{setFplot_dict}}.}

\item{mod.title}{Character scalar. The title of the legend, in case there is a moderator. By default it is the moderator name modified by dict if the moderator is numeric (otherwise default is empty). To display no title, set it to \code{NULL}.}

\item{labels.angle}{Only if the labels of the x-axis are tilted. The angle of the tilt.}

\item{cex.axis}{Cex value to be passed to biased labels. By defaults, it finds automatically the right value.}

\item{trunc}{If the main variable is a character, its values are truncaded to \code{trunc} characters. Default is 20. You can set the truncation method with the argument \code{trunc.method}.}

\item{trunc.method}{If the elements of the x-axis need to be truncated, this is the truncation method. It can be "auto", "trimRight" or "trimRight".}

\item{...}{Other elements to be passed to plot.}
}
\description{
This function plots distributions of items (a bit like an histogram) which can be easily conditioned over.
}
\examples{

# generating data
y = rnorm(200, mean=10)
x = round(y)
id = sample(c("XX", "XY"), 200, TRUE)

base = data.frame(y, x, id)
base$x[base$id == "XY"] = base$x[base$id == "XY"] + 2

# Frequency of y => classic histogram
plot_distr(base$y)

# Frequency of x => different layout because made of integers
plot_distr(base$x)

#
# Now splitting wrt id

# default behavior depends on the data. Here not sorted:
plot_distr(x~id, base)

# now sorted
plot_distr(x~id, base, maxFirst = TRUE)

# In log
plot_distr(x~id, base, toLog = TRUE)

#
# Using iris data
plot_distr(iris$Species)

plot_distr(Species~1|Sepal.Width, iris, onTop="none")

plot_distr(Sepal.Width~1, iris)



}
