% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscfuns.R
\docType{data}
\name{us_pub_econ}
\alias{us_pub_econ}
\title{Publication data sample}
\format{
\code{us_pub_econ} is a data table with 30,756 observations and 6 variables.

\itemize{
\item{paper_id: Numeric identifier of the publication.}
\item{year: Year of publication.}
\item{institution: Institution of the authors of the publication.}
\item{journal: Journal/conference name.}
\item{jnl_top_25p: 0/1 variable of whether the journal belongs to the top 25\% in terms of average cites.}
\item{jnl_top_5p: 0/1 variable of whether the journal belongs to the top 5\% in terms of average cites.}

}
}
\source{
The source is Microsoft Academic Graph (see reference).
}
\usage{
data(us_pub_econ)
}
\description{
This data reports the publications of U.S. institutions in the field of economics between 1985 and 1990.
}
\references{
Arnab Sinha, Zhihong Shen, Yang Song, Hao Ma, Darrin Eide, Bo-June (Paul) Hsu, and Kuansan Wang. 2015. An Overview of Microsoft Academic Service (MAS) and Applications. In Proceedings of the 24th International Conference on World Wide Web (WWW '15 Companion). ACM, New York, NY, USA, 243-246.
}
\keyword{datasets}
